--- src/components/Ingredient/ingredient.types.ts
@@ -1,5 +1,0 @@
-import { IRecipeIngredientComplete } from "../../Pages/Recipe/recipe.types";
-
-export interface IngredientProps {
-  ingredients: IRecipeIngredientComplete[];
-}
