--- src/components/RecipeCard/recipeCard.types.ts
@@ -1,7 +1,0 @@
-import { IRecipeBlob } from "../../API/API.types";
-
-export interface IRecipeCardProps {
-  recipeCardData: IRecipeBlob;
-  ingredientCount: number;
-  key: string | number;
-}
