--- src/components/RecipeCardList/recipeCardList.types.ts
@@ -1,6 +1,0 @@
-import { IRecipeBlob } from "../../API/API.types";
-
-export interface IRecipeCardListProps {
-  recipeCardDataList: IRecipeBlob[];
-  ingredientCount: number;
-}
