+++ src/components/RecipeInfo/recipeInfo.tsx
@@ -1,14 +1,18 @@
 import { Container, Grid } from "@mui/material";
 import { Box } from "@mui/system";
-import { IRecipeInfoProps } from "./recipeInfo.types";
 import { RecipeStat } from "../RecipStats/recipeStats";
 import { StyledRecipeBox } from "./recipeInfo.styles";
 
-export const RecipeInfo: React.FC<IRecipeInfoProps> = ({
+export const RecipeInfo = ({
   portions,
   cookingTime,
   ingredientMatches,
   inputIngredient,
+}: {
+  portions: number;
+  cookingTime: number;
+  ingredientMatches: number;
+  inputIngredient: number;
 }) => {
   return (
     <StyledRecipeBox>
