--- src/components/RecipeInfo/recipeInfo.types.ts
@@ -1,6 +1,0 @@
-export interface IRecipeInfoProps {
-  portions: number;
-  cookingTime: number;
-  ingredientMatches: number;
-  inputIngredient: number;
-}
