+++ src/components/Steps/steps.tsx
@@ -1,10 +1,18 @@
 import { Box, Stack } from "@mui/system";
 import { useEffect, useState } from "react";
-import { IStepsProps, IRecipeStepsDoneProps } from "./steps.types";
 import { StyledH2 } from "../../styles/theme";
 import { Step, StepContainer } from "./steps.styles";
+import { IRecipeSteps } from "../../API/API.types";
 
-export const Steps: React.FC<IStepsProps> = ({ recipeSteps }) => {
+export interface IRecipeStepsDoneProps extends IRecipeSteps {
+  done?: boolean;
+}
+
+export const Steps = ({
+  recipeSteps,
+}: {
+  recipeSteps: IRecipeStepsDoneProps[];
+}) => {
   const [steps, setSteps] = useState<IRecipeStepsDoneProps[]>([]);
 
   useEffect(() => {
