--- src/components/Steps/steps.types.ts
@@ -1,9 +1,0 @@
-import { IRecipeSteps } from "../../API/API.types";
-
-export interface IRecipeStepsDoneProps extends IRecipeSteps {
-  done?: boolean;
-}
-
-export interface IStepsProps {
-  recipeSteps: IRecipeStepsDoneProps[];
-}
